move texture caching code to material system (fixes #1178)
